requirements.txt
pytest~=8.3.3
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client~=4.4.0
selenium~=4.27.1
pydantic_settings~=2.7.1
requests~=2.32.3
python-dotenv~=1.0.1
